MAPPED ROUTE B100 from SQL query
There are six coordinates that I had to manually order. Otherwise the route was nonsensical.
I have a feeling that these six coordiantes were the six that took us from 21 stops to 27 stops
on a previous query.

Theis route is partially validated with googles suggestion for a route (from stop 1 to stop last).
https://www.google.com/maps/dir/40.608531,-73.957395/40.606791,+-73.916037/@40.6105346,-73.9328786,15z/am=t/data=!4m7!4m6!1m0!1m3!2m2!1d-73.916037!2d40.606791!3e3